Server refactoring (#8)
* Added system console setting to disable certificate validation
* Fixed the web app blank screen issue and updated server manifest
* Updated configuration.go
* Refactored API routes to a separate file.
