+++ .golangci.yml
@@ -52,6 +52,8 @@ issues:
     - path: server/configuration.go
       linters:
         - unused
+    - path: server/utils.go
+      linters:
         - gosec # G402: TLS InsecureSkipVerify set true.
     - path: _test\.go
       linters:
