+++ go.mod
@@ -1,9 +1,10 @@
 module github.com/CollaboraOnline/collabora-mattermost
 
-go 1.12
+go 1.16
 
 require (
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/mattermost/mattermost-server/v5 v5.34.2
 	github.com/pkg/errors v0.9.1
+	github.com/stretchr/testify v1.7.0
 )
